+++ src/discord/ModalSubmitHandlers.ts
@@ -1,0 +1,12 @@
+import {ModalConfig} from "./ModalConfig";
+import {PineconeModal} from "./handlers/modals/PineconeModal";
+import {EmbedLimitModal} from "./handlers/modals/EmbedLimitModal";
+import {TokenLimitsModal} from "./handlers/modals/TokenLimitsModal";
+import {OpenAIAPIKeyModal} from "./handlers/modals/OpenAIAPIKeyModal";
+
+export const ModalSubmitHandlers: ModalConfig[] = [
+    PineconeModal,
+    EmbedLimitModal,
+    TokenLimitsModal,
+    OpenAIAPIKeyModal,
+];
