+++ src/discord/handlers/buttons/TokenLimitsButtonHandler.ts
@@ -1,0 +1,15 @@
+import {ButtonCommand} from "../../ButtonCommand";
+import {TokenLimitsModal} from "../modals/TokenLimitsModal";
+
+export const TokenLimitsButtonHandler: ButtonCommand = {
+    id: 'SET-TOKEN-LIMITS',
+    async run(client, buttonInteraction) {
+        await TokenLimitsModal.show(buttonInteraction);
+
+        await buttonInteraction.editReply({
+            content: 'Updating Token Limits...',
+            components: [],
+            embeds: [],
+        });
+    }
+}
